+++ simple-jwt-login/views/auth-view.php
@@ -85,6 +85,7 @@ if (! defined('ABSPATH')) {
             Parameters:<br />
             <b>email</b> -> <?php echo __('to login with email', 'simple-jwt-login');?><br />
             <b>username</b> -> <?php echo __('to login with username', 'simple-jwt-login');?><br />
+            <b>login</b> -> <?php echo __('to login with username or email', 'simple-jwt-login');?><br />
             <b>password</b> -> <?php echo __('your password', 'simple-jwt-login');?><br />
             <b>password_hash</b> -> <?php echo __('your hashed password from the database', 'simple-jwt-login');?><br />
         </p>
@@ -106,6 +107,30 @@ if (! defined('ABSPATH')) {
                 </button>
             </span>
         </div>
+        
+        <div>
+            <b><?php echo __('OR', 'simple-jwt-login');?></b>: 
+        </div>
+        
+        <div class="generated-code">
+            <span class="method">POST:</span>
+            <span class="code">
+                <?php
+                $sampleUrlParams = [
+                    'username'    => __('Username', 'simple-jwt-login'),
+                    'password' => __('Password', 'simple-jwt-login')
+                ];
+
+                echo esc_html($jwtSettings->generateExampleLink(RouteService::AUTHENTICATION_ROUTE, $sampleUrlParams));
+                ?>
+            </span>
+            <span class="copy-button">
+                <button class="btn btn-secondary btn-xs">
+                    <?php echo __('Copy', 'simple-jwt-login'); ?>
+                </button>
+            </span>
+        </div>
+
         <div>
             <b><?php echo __('OR', 'simple-jwt-login');?></b>: 
         </div>
@@ -114,7 +139,7 @@ if (! defined('ABSPATH')) {
             <span class="code">
                 <?php
                 $sampleUrlParams = [
-                    'username'    => __('Username', 'simple-jwt-login'),
+                    'login'    => __('Username_or_email', 'simple-jwt-login'),
                     'password' => __('Password', 'simple-jwt-login')
                 ];